Moving MultiMapDataStore to mapsforge-map library. #668
