+++ Applications/Android/Samples/src/main/java/org/mapsforge/applications/android/samples/MultiMapLowResWorld.java
@@ -19,7 +19,7 @@ import android.os.Bundle;
 
 import org.mapsforge.map.datastore.MapDataStore;
 import org.mapsforge.map.reader.MapFile;
-import org.mapsforge.map.reader.MultiMapDataStore;
+import org.mapsforge.map.datastore.MultiMapDataStore;
 
 import java.io.File;
